% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_saved_list.R
\name{export_saved_list}
\alias{export_saved_list}
\title{export_saved_list}
\usage{
export_saved_list(listid, listname = NULL)
}
\arguments{
\item{listid}{Numeric saved list identifier}

\item{listname}{Name for vector of saved list in global environment}
}
\value{
A vector of all VANIDs in the saved list
}
\description{
Function to pull all VANIDs from a given saved list.
#'
}
