% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_van_event_url.R
\name{generate_van_event_url}
\alias{generate_van_event_url}
\title{generate_van_event_url}
\usage{
generate_van_event_url(event_id)
}
\arguments{
\item{event_id}{Numeric event identifier}
}
\value{
A URL for the event
}
\description{
Determines the URL link for a given eventID so it's easier to find it in the smartVAN UI. Because why would they just let us search by event ID?? Can't let things be too easy! This process can work for many other ID types so this function could be expanded in the future.
}
