% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/van_helper_functions.R
\name{van_auth}
\alias{van_auth}
\title{van_auth}
\usage{
van_auth(username, password, db)
}
\arguments{
\item{username}{username associated with API Key}

\item{password}{The API Key}

\item{db}{For db, enter 1 for EA, 0 for myVoters}
}
\description{
Function to turn API key information in proper format for data retrieval
}
